pyproject.toml
[project]
name = "rl-discord-scoreworm"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.7.1",
    "matplotlib>=3.10.8",
    "mss>=10.1.0",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
